fix(material): D-413 打通小数链路——清掉到货/采购/扫码入库入口的取整
D-410 只改了实体与 DB 字段（编译层面通过），但入口与前端仍在取整，
1.32 米照旧被截成 1，P0 实际未生效。本批把这些最后一公里的截断点清掉。

后端：
- MaterialPurchaseOrchestrator.updateArrivedQuantity（/update-arrived-quantity 主接口）
  原用 helper.coerceInt（intValue 截断），改用现成的 coerceBigDecimal（setScale(4,HALF_UP)）；
  采购数量也不再 intValue()，到货比例判定改用 BigDecimal 比较。
- createInstruction：采购数量同样改用 coerceBigDecimal
  （purchase_quantity 早就是 DECIMAL，这里是长期瓶颈）
- MaterialStockController.scan-inbound：新增 parseQuantity() 走字符串解析，
  不用 doubleValue 避免精度毛刺；scanInbound 签名改 BigDecimal
- scan-outbound 保持整数：MaterialOutboundLog.quantity 未纳入迁移，需另开迁移

前端：
- 到货数量输入框两处（PurchaseActionModals / InlinePurchasePanel）原为
  min={1} step={1} precision={0} 锁死整数，且提示文案还写着
  「小数到货需数据模型升级后开放」（已过时）→ 统一为 0.01/0.01/2
- ConfirmCompleteModal 同上，并把写死的单位「件」改为取物料单位
- 预填到货量去掉 Math.max(1, Math.round(...))：
  剩余 0.5 时会预填 1，反而超过剩余量被后端「到货数量超出采购数量」拒绝
- 领料去掉 Math.floor：剩余 0.5 会算成 0 → pickQty>0 为假 → 点了没反应（静默失效）
  改传精确值，由 handleWarehousePick 判定；不足 1 时给出明确提示
  （领料单 quantity 仍是 Integer，小数领料需另开迁移）

+++ backend/src/main/java/com/fashion/supplychain/warehouse/orchestration/MaterialWarehouseOperationOrchestrator.java
@@ -414,7 +414,8 @@ public class MaterialWarehouseOperationOrchestrator {
     }
 
     @Transactional(rollbackFor = Exception.class)
-    public MaterialStock scanInbound(String materialCode, int quantity, String warehouseLocation,
+    /** D-410 收尾：扫码入库数量支持小数（int 会把 1.32 米截断成 1） */
+    public MaterialStock scanInbound(String materialCode, BigDecimal quantity, String warehouseLocation,
                                       String warehouseAreaId, String sourceType, String remark,
                                       String materialName, String materialType, String color, String size) {
         TenantAssert.assertTenantContext();
